chainConfig
