+++ packages/backend/src/modules/tvs/tools/LocalExecutor.ts
@@ -1,5 +1,5 @@
 import type { Env, Logger } from '@l2beat/backend-tools'
-import { type ChainConfig, ProjectService } from '@l2beat/config'
+import type { ChainConfig, ProjectService } from '@l2beat/config'
 import {
   BlockProvider,
   CoingeckoClient,
@@ -37,10 +37,7 @@ export class LocalExecutor {
     timestamps: UnixTime[],
     latestMode: boolean,
   ): Promise<TokenValue[]> {
-    const chainConfigs = await new ProjectService().getProjects({
-      select: ['chainConfig'],
-    })
-    const { prices, amounts } = extractPricesAndAmounts(config, chainConfigs)
+    const { prices, amounts } = extractPricesAndAmounts(config)
 
     const dataFormulaExecutor = await this.initDataFormulaExecutor(amounts)
